changed transforms

+++ src/utils/transforms.py
@@ -43,4 +43,4 @@ class Normalize(object):
 class ToTensor:
     def __call__(self, pointcloud):
         assert len(pointcloud.shape) == 2
-        return torch.from_numpy(pointcloud)
+        return torch.from_numpy(pointcloud.astype(np.float32))
